ml-deployment/requirements.txt
flask==3.0.2
torch
torchvision
pillow
numpy
gunicorn
requests
fastapi
uvicorn
python-multipart
